consolidate duplicate parser
