--- pkg/performer/career_test.go
@@ -1,48 +1,0 @@
-package performer
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-)
-
-func TestParseCareerLength(t *testing.T) {
-	intPtr := func(v int) *int { return &v }
-
-	tests := []struct {
-		name      string
-		input     string
-		wantStart *int
-		wantEnd   *int
-		wantErr   bool
-	}{
-		{"single year", "2005", intPtr(2005), nil, false},
-		{"year range with spaces", "2005 - 2010", intPtr(2005), intPtr(2010), false},
-		{"year range no spaces", "2005-2010", intPtr(2005), intPtr(2010), false},
-		{"year dash open", "2005 -", intPtr(2005), nil, false},
-		{"year dash open no space", "2005-", intPtr(2005), nil, false},
-		{"dash year", "- 2010", nil, intPtr(2010), false},
-		{"year present", "2005-present", intPtr(2005), nil, false},
-		{"year Present caps", "2005 - Present", intPtr(2005), nil, false},
-		{"whitespace padding", "  2005 - 2010  ", intPtr(2005), intPtr(2010), false},
-		{"empty string", "", nil, nil, true},
-		{"garbage", "not a year", nil, nil, true},
-		{"partial garbage start", "abc - 2010", nil, nil, true},
-		{"partial garbage end", "2005 - abc", nil, nil, true},
-		{"year out of range", "1800", nil, nil, true},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			start, end, err := ParseCareerLength(tt.input)
-			if tt.wantErr {
-				require.Error(t, err)
-				return
-			}
-			require.NoError(t, err)
-			assert.Equal(t, tt.wantStart, start)
-			assert.Equal(t, tt.wantEnd, end)
-		})
-	}
-}
